pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "paho-mqtt>=2.1.0",
    "psycopg2>=2.9.10",
    "pydantic>=2.10.4",
    "pymongo>=4.10.1",
    "python-decouple>=3.8",
    "pytz>=2024.2",
    "redis>=5.2.1",
    "uvicorn[standard]>=0.34.0",
]
